src/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.2
PyMySQL==1.1.1
python-dotenv==1.0.1
pytest==8.2.2
pandas==2.2.2
openpyxl==3.1.5
Flask-Migrate==4.0.7
python-docx==1.1.0
reportlab==4.2.0